+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * fix division on single argument (now `(/ n)` == `(/ 1 n)`)
 * fix complex operation that result in real (e.g. multiplication over conjugation)
 * fix `list-ref` according to R7RS errata
+* fix formatter (pretty print) on multiline strings and atoms
 
 ## 1.0.0-beta.11
 ### Breaking